Replace random search with list of positive words
Starting the app for the first time would trigger a random search which
can result in negative searches. Therefore a list of positive words has
been created to replace that.
Also, added "random search", which takes one of the positive words
randomly, when leaving the search field empty and
starting the search.